setup permission + gates

+++ laravel-admin/app/Http/Controllers/RoleController.php
@@ -2,8 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use DB;
 use App\Models\Role;
 use Illuminate\Http\Request;
+use App\Http\Resources\RoleResource;
+use Illuminate\Support\Facades\Gate;
 use Symfony\Component\HttpFoundation\Response;
 
 class RoleController extends Controller
@@ -11,7 +14,8 @@ class RoleController extends Controller
 
     public function index()
     {
-        return Role::all();
+        Gate::authorize('view', 'roles');
+        return RoleResource::collection(Role::all());
     }
 
     /**
@@ -22,9 +26,18 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('edit', 'roles');
         $role = Role::create($request->only('name'));
+        if ($permissions = $request->input(('permissions'))) {
+            foreach ($permissions as $permission_id) {
+                DB::table('role_permission')->insert([
+                    'role_id' => $role->id,
+                    'permission_id' => $permission_id,
 
-        return response($role, Response::HTTP_CREATED);
+                ]);
+            }
+        }
+        return response(new RoleResource($role), Response::HTTP_CREATED);
     }
 
     /**
@@ -35,7 +48,8 @@ class RoleController extends Controller
      */
     public function show($id)
     {
-        return  Role::find($id);
+        Gate::authorize('view', 'roles');
+        return new RoleResource(Role::find($id));
     }
 
     /**
@@ -47,15 +61,36 @@ class RoleController extends Controller
      */
     public function update(Request $request, $id)
     {
+        Gate::authorize('edit', 'roles');
         $role = Role::find($id);
+
         $role->update($request->only('name'));
-        return response($role, Response::HTTP_ACCEPTED);
+
+        Db::table('role_permission')->where('role_id', $role->id)->delete();
+
+
+
+        if ($permissions = $request->input(('permissions'))) {
+            foreach ($permissions as $permission_id) {
+                DB::table('role_permission')->insert([
+                    'role_id' => $role->id,
+                    'permission_id' => $permission_id,
+
+                ]);
+            }
+        }
+
+        return response(new RoleResource($role), Response::HTTP_ACCEPTED);
     }
 
 
     public function destroy($id)
     {
+        Gate::authorize('edit', 'roles');
+        Db::table('role_permission')->where('role_id', $id)->delete();
+
         Role::destroy($id);
+
         return response(null, Response::HTTP_NO_CONTENT);
     }
 }
